Add files via upload
The images depict two separate instances of a password strength checker script written in Python, being run in a Google Colab environment. The script checks a password against a set of rules, including the presence of special characters and a list of weak passwords.

In the first image, the password "JAMESBond123@" is tested and receives a score of 6/6, which is classified as a "Strong password!".

In the second image, the password "1234567890" is tested. It receives a lower score of 3/6 and the program provides suggestions for improvement, such as making it at least 12 characters long and adding uppercase and lowercase letters.

The third image shows a GitHub interface where a user is preparing to upload a file named Query successful.pdf to a repository called Password-Strength-Validator. The user has the option to commit the changes directly to the main branch.